#50 GET de uma task criado.

+++ server/core/views.py
@@ -22,7 +22,8 @@ def tasklists(request):
         task_list = list(tasks)  
         return JsonResponse(task_list, safe=False)
     else:
-        return HttpResponse("POST")    
+        return HttpResponse("POST") 
+
 @csrf_exempt
 @require_http_methods(["PUT","DELETE","GET"])
 def tasklistsid(request, tasklist_id):
@@ -66,25 +67,32 @@ def tasklists_id_tasks(request, tasklist_id):
 
         try: 
             task.save()  
-        except Exception as screen_error:
-            logging.error(screen_error)
+        except Exception as error:
+            logging.error(error)
             return JsonResponse({'message': 'Error when saving'}, status=500) 
         
         return JsonResponse({"message": 'Task created'}, status=201)
 
 
 @csrf_exempt
-@require_http_methods(["DELETE"])
-def tasks_id(request, task_id):
-    try:
-        task = Task.objects.get(id=task_id)
-        task.delete()
-    except Task.DoesNotExist:
-        return JsonResponse({'message': ' Task not found.'}, status=404) 
+@require_http_methods(["GET", "DELETE"])
+def task_id(request, task_id):
+    if request.method == "GET":
+        try:
+            task = Task.objects.get(id=task_id)
+            single_task = model_to_dict(task)
+        except Task.DoesNotExist: 
+            return JsonResponse({'message': ' Task not found.'}, status=404)    
+        
+        return JsonResponse(single_task, safe=False)      
+    else:    
+        try:
+            task = Task.objects.get(id=task_id)
+            task.delete()
+        except Task.DoesNotExist:
+            return JsonResponse({'message': ' Task not found.'}, status=404) 
 
-    return JsonResponse({"message": 'Task deleted'}, status=200)
-
-    
+        return JsonResponse({"message": 'Task deleted'}, status=200)      
 
 
 # Tasklists
@@ -96,9 +104,11 @@ def tasks_id(request, task_id):
 
 # Tasks
 # GET /tasklists/:id/tasks/ - retorna todas as tasks de uma determinada tasklist ✅
+# GET /tasklists/:id/tasks/:id - retorna apenas uma task dentre as tasks de uma determinada tasklist 
 # POST /tasklists/:id/tasks/ - cria uma nova task✅
 # PUT/PATCH? /tasks/:id/ - atualiza apenas uma task de uma determinada tasklist
 # DELETE /tasks/:id/ - deleta apenas uma task✅
+
